# storage/api-docs/api-docs.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "BasePHP API",
...
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API REST versionada v1"
        },
        {
            "url": "/",
            "description": "Raiz da aplicação, usada por endpoints auxiliares como OAuth2 Passport"
        }
    ],
    "paths": {
...
        "/api/v1/auth/change-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Altera a senha do usuário autenticado.",
                "description": "Endpoint protegido por OAuth2. Usado quando o usuário precisa trocar a senha antes de acessar recursos tenant-aware.",
                "operationId": "changePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha-atual"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NovaSenhaForte@123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NovaSenhaForte@123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha alterada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido ou ausente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "user.password.change"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiSuccessResponse": {
                "description": "Envelope padrão para respostas de sucesso.",
                "required": [
                    "success",
                    "message",
                    "data"
                ],
                "properties": {
                    "success": {
                        "description": "Indica se a operação foi concluída com sucesso.",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Mensagem padronizada para o consumidor da API.",
                        "type": "string",
                        "example": "Operação realizada com sucesso."
                    },
                    "data": {
                        "description": "Objeto de retorno da operação.",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApiErrorResponse": {
                "description": "Envelope padrão para respostas de erro.",
                "required": [
                    "success",
                    "message",
                    "errors"
                ],
                "properties": {
                    "success": {
                        "description": "Indica que a operação falhou.",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Mensagem principal do erro.",
                        "type": "string",
                        "example": "Erro ao processar a requisição."
                    },
                    "errors": {
                        "description": "Lista de erros detalhados, quando aplicável.",
                        "type": "array",
                        "items": {
                            "properties": {
                                "property": {
                                    "description": "Campo relacionado ao erro, quando aplicável.",
                                    "type": "string",
                                    "example": "email"
                                },
                                "type": {
                                    "description": "Tipo técnico ou funcional do erro.",
                                    "type": "string",
                                    "example": "VALIDATION"
                                },
                                "message": {
                                    "description": "Descrição detalhada do erro.",
                                    "type": "string",
                                    "example": "O campo email é obrigatório."
                                },
                                "exception": {
                                    "description": "Classe da exceção tratada, quando exposta de forma segura.",
                                    "type": "string",
                                    "example": "ValidationException"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Autenticação OAuth2 via Laravel Passport usando Authorization Code + PKCE para usuários humanos e Client Credentials para integrações sistema-a-sistema.",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/oauth/token",
                        "scopes": {
                            "admin.full": "Acesso administrativo total",
                            "tenant.access": "Acesso ao tenant autenticado",
                            "user.profile": "Acesso ao perfil autenticado",
                            "user.password.change": "Alteração da própria senha",
                            "tenant.users.read": "Listagem de usuários do tenant",
                            "tenant.users.write": "Criação e manutenção de usuários do tenant",
                            "tenants.read": "Listagem de tenants",
                            "tenants.write": "Criação e manutenção de tenants",
                            "users.read": "Listagem de usuários",
                            "users.write": "Criação e manutenção de usuários",
                            "queues.read": "Consulta operacional de filas",
                            "queues.write": "Ações operacionais em filas",
                            "emails.read": "Consulta de envios de e-mail",
                            "emails.write": "Envio e reprocessamento de e-mails"
                        }
                    },
                    "clientCredentials": {
                        "tokenUrl": "http://localhost/oauth/token",
                        "scopes": {
                            "system.health": "Verificação operacional sistema-a-sistema"
                        }
                    }
                }
            },
            "tenantHeader": {
                "type": "apiKey",
                "description": "Código do tenant ativo. Exemplo: tenant-main.",
                "name": "X-Tenant-Id",
                "in": "header"
            }
        }
    },
...
}